rootmap: correctly join paths when looking for physical root
To iterate faster, it's convenient to be able to run `rdcore rootmap`
directly from the real root (mounted over from the development host)
and pointing it at `/`. But the naive appending we do here to find the
physical root would give `//sysroot` which isn't in the mount table
(obviously we could also canonicalize, but it's just cleaner to join it
correctly in the first place).