test/images/: change some filenames
With the OSBuild live ISO/PXE creation a few of the filenames
changed sligthly. The kernel doesn't have `-` after the `kernel`
word and the ISO has `live-iso` in the filename now. Let's
change that and also add `live-` to the front because each of
the artifacts is prepended with that.