Cargo.toml: stop setting debug=true for release profile
The related comment here:

> # We assume we're being delivered via e.g. RPM which supports split debuginfo

I don't think is really needed because when building the RPM the
%cargo_build macro expands to:

```
rust-srpm-macros-26.3-3.fc41.noarch
cargo-rpm-macros-26.3-3.fc41.noarch
/usr/bin/env CARGO_HOME=.cargo RUSTC_BOOTSTRAP=1 RUSTFLAGS='-Copt-level=3 -Cdebuginfo=2 -Ccodegen-units=1 -Cstrip=none --cap-lints=warn' /usr/bin/cargo build -j${RPM_BUILD_NCPUS} -Z avoid-dev-deps --profile rpm
```

Which clearly uses the `rpm` profile, which does inherit from the
release profile, but also sets `debug` separately [1]. The macro
also has `-Cdebuginfo=2` so that should do it too.

Let's just drop it here so that when developers build and
`--target release` it does what they expect and creates a smaller
binary.

[1] https://pagure.io/fedora-rust/rust-packaging/blob/1402e757e3200e6f06b8a9c0db6350de8473de73/f/macros.d/macros.cargo#_86